+++ FreePIE.Tests.Core/FreePIE.Tests.Core.csproj
@@ -55,6 +55,7 @@
     </CodeAnalysisDependentAssemblyPaths>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CurveHelperTest.cs" />
     <Compile Include="FilterHelperTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PluginInvokerTests.cs" />
